fix: Add publisher to member offerings and extract profile routes (#796)
- Add 'publisher' to VALID_MEMBER_OFFERINGS to match MemberOffering type
- Create centralized VALID_* constants for runtime validation to prevent type drift:
  - VALID_MEMBER_OFFERINGS, VALID_ORGANIZATION_ROLES, VALID_ASSIGNABLE_ROLES
  - VALID_LEGAL_DOCUMENT_TYPES, VALID_REVENUE_TIERS
- Extract member profile routes from http.ts to routes/member-profiles.ts (~530 lines)
- Add 'trialing' to subscription status types for trial user support
- Fix slug validation to properly check reserved keywords

Closes #796
